knowledge-refresh-pipeline/requirements.txt
# Knowledge Refresh Pipeline Requirements
# Core dependencies for automated knowledge refresh cycles

# Async and concurrency
asyncio==3.4.3
aiohttp==3.9.1
aiofiles==23.2.1

# Data processing and storage
redis==5.0.1
chromadb==0.4.18
qdrant-client==1.7.0

# Vector embeddings and NLP
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
numpy==1.24.4

# Web scraping and content extraction
beautifulsoup4==4.12.2
feedparser==6.0.10
requests==2.31.0
lxml==4.9.3

# Text analysis and quality assessment
textstat==0.7.3
nltk==3.8.1

# Configuration and serialization
PyYAML==6.0.1
pydantic==2.5.2
python-dotenv==1.0.0

# Monitoring and logging
structlog==23.2.0

# HTTP client enhancements
httpx==0.25.2

# Data validation and parsing
marshmallow==3.20.2

# Scheduling and time handling
APScheduler==3.10.4
pytz==2023.3

# Database utilities
sqlalchemy==2.0.23

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional dependencies for specific features
# Uncomment as needed:

# For advanced NLP features
# spacy==3.7.2

# For cloud storage integration
# boto3==1.34.0  # AWS
# google-cloud-storage==2.10.0  # Google Cloud
# azure-storage-blob==12.19.0  # Azure

# For database backends
# psycopg2-binary==2.9.9  # PostgreSQL
# pymongo==4.6.1  # MongoDB

# For notification systems
# slack-sdk==3.25.0
# sendgrid==6.11.0

# For metrics and observability
# prometheus-client==0.19.0
# grafana-api==1.0.3